Module Middleware.Diverter

type (!'next_input, !'next_output, !'output) t = ('next_input, 'next_output, 'output) Make(Monads.Identity).Diverter.t =

| Stop of 'output0 Monads.Identity.t
(*
Control stops here, and the 'output is returned to the wrapping Middleware.
*)

| Continue of 'next_input0 * ('next_output0 -> 'output1 Monads.Identity.t)
(*
Control continues to the next Middleware g, which receives 'next_input as input. After g runs, its result is transformed with the given function.
*)

Union, determining whether or not to run a subsequent Middleware or return early.